Avoid early initializers in the compiler.
Early initializers are deprecated in Scala 2.13. Unfortunately, the
compiler API is still designed in a fashion that requires early
initialization of the `val global`s. We use a trick to provide that
without early initializers: provide the `val global` as a class
constructor parameter, which are still initialized early.

This commit finalizes the confinement of warnings in the compiler
to `Compat210Component`, including with the latest nightly of
Scala 2.13.