Use a dedicated `InlineableClassStructure` in the optimizer.
We previously used a `RecordValue`/`RecordType` as proxy for the
information about an inlineable class that the optimizer needed to
carry. This commit instead uses a dedicated
`InlineableClassStructure`, which abstracts over some details,
notably what constitutes the identity of a field, and how that
identity is mapped to a record field.