Detach VirtualScalaJSIRFile from other virtual files
This will make it easier to make VirtualScalaJSIRFile asynchronous,
since we do not have an exposed async I/O abstraction anymore.

Further, note that a `version` is only needed for
`VirtualScalaJSIRFile`, but not for other virtual files. Also, note
that we do not allow a version to change anymore over the lifetime of
a `VirtualScalaJSIRFile`; this is currently always true in practice
and simply makes the equivalent requirement of `IRFileCache`
explicit.