Rename JS* IR nodes involving selections.
* `JSDotSelect` -> `JSPrivateSelect`
* `JSBracketSelect` -> `JSSelect`
* `JSBracketMethodApply` -> `JSMethodApply`
* `JSSuperBracketSelect` -> `JSSuperSelect`
* `JSSuperBracketCall` -> `JSSuperCall`

The new names are more semantically meaningful, and correspond
to how those nodes are used.

`JSDotMethodApply` was removed because it was unused.

`JSDelete` was never used with `JSDotSelect`s, so it was
specialized for the semantics with `JSBracketSelect`s (now
`JSSelect`s) by directly integrating the qualifier and the item.