Do not exclude bridges when looking for JUnit-related methods.
It turns out that this was only necessary for one very peculiar
test method, which was defined as

    @Test def toString()(): Unit = ...

The double-`()` allowed it not to clash with the `toString()`
inherited from `Object`, but then for some reason erasure was
generating a bridge (although it shouldn't).

Instead of having the JUnit plugin working around that single
method, we rename it.

Not excluding bridges is more regular with respect to how JUnit
works on the JVM.