Fix #3538: Ignore mixin forwarders when looking for export alternatives.
Mixin forwarders are cloned with a type history that does not go
far enough back in time to see varargs. By skipping them when
looking for the alternatives of an export, we look further in the
class hierarchy to the original definition in the trait, which does
expose the entire type history that we need.