Use `getPackageObject` instead of `getPackage` to fix a deprecation.
`getPackage(TermName)` was deprecated upstream in
https://github.com/scala/scala/commit/6bcf47d364b77da340fe347893f1a8e13d457589
in favor of a new `getPackage(String)`. We cannot use the new
method because it is not available in older versions.

However we can use `getPackageObject`, which is more appropriate
anyway, and which we already used to get other package objects like
`runtime` and `special`.