Fix the overriding-pairs `Cursor` used by PrepJSInterop for 2.13.
In the upstream commit
https://github.com/scala/scala/commit/fde215f606dc63813bea3b6bc64176aa551b40ce,
a new abstract method `skipOwnerPair` was introduced in `Cursor`.
We now implement it in the same way as `RefChecks` does, since
previously we used exactly the same `Cursor` as `RefChecks`. There
is some compat code to allow the new code to compile with older
Scala versions.