[no-master] Upgrade to Google Closure Compiler v20181210.
Backport of 2ad657c6c86001e29c0228f6f034a185ffd3cf97.

Because of too much lockdown in the ErrorManager API of Closure,
we cannot send the errors and warnings to our `logger` anymore.
Instead, they will be sent to the standard error, if any. Note that
in theory, this never happens.

This commit fixes #2815 as a side effect: the new version of GCC
does not exhibit the bug that caused that issue.