Support the new flat patmat ASTs of Scala 2.12.9.
See https://github.com/scala/scala/pull/8061 upstream.

It is now possible for jumps to the next-case label to be buried in
arbitrary positions within a case body. We support that by
enclosing a case body with a labeled block, in case at least one
`return` to the label could not be caught upstream.

Doing so produces significantly worse code for cases that only have
one test (and hence one `return`), which are by far the most common
scenario. Therefore, we add an optimization to rewrite cases with a
single `return` of the case to the previous style of `If`. This in
turn allows the optimization of the `matchEnd` label to kick in,
and production of better code.

When there are more than one `return`, the new flat structure is
beneficial for our IR as much as for scalac Trees, and therefore we
keep it as is.