Fix #2629: Review the encoding of fields in the IR.
Instead of being name-mangled in the IR, their identity now
explicitly refers to their declaring class. The mangling is
deferred to the emitter.

This increases the code size because the mangling done by the
emitter includes the full name of the declaring class, whereas
previously we used the number of superclasses in the hierarchy.
The new mangling is easier to deal with in incremental runs of the
linker. A more efficient stategy could be reintroduced in the
future, without breaking compatibility.