Always require source map support, even in ES2015.
Historically, when emitting ES2015 code, we did not require source
maps, and consequently did not require the source map tests to
pass. This exclusion can be traced back all the way to when we
first started testing the ES2015 output in the CI in
f69dd97b3fe8144f8ead0bdef7e9ba6caccb2a96. I suspect that this was
done because source map support in ES2015 was not good enough in
engines and/or in `source-map-support`.

Now, however, it seems that the source map tests are happy to pass
with the ES2015 output, so we unconditionally enable them by
default.