Fix #3668: Split the test bridge into a separate artifact.
The scalajs-test-interface artifact actually contained two
independent concerns:

* The Scala.js version of the `sbt.testing.*` APIs, and
* The JS-side implementation of the test adapter/bridge protocol.

Concrete testing frameworks depend on the former, but need not
know about the latter. Moreover, the protocol in the bridge needs
to be in sync with the one in the adapter.

Keeping those two concerns together meant that transitive
dependencies of testing frameworks could force a version of the
bridge that was newer than the adapter. This possibility forced us
to lock down the protocol in very serious ways.

This commit splits those two concerns in two different artifacts,
so that we can evolve the protocol more freely.

For compatibility, `ScalaJSClassLoader` and `TestUtils` are kept
in the scalajs-test-interface artifact. They are both already
deprecated, so they will be gone in Scala.js 1.x. The classes that
were moved to scalajs-test-bridge are package-private, so there is
no issue for them.