Fix #3686: Add `js.import(specifier)` to support ES dynamic imports.
Tests are rudimentary, because full testing requires asynchronous
tests, which are not available in 0.6.x. Full tests should be
added in the 1.x series.

The changes in the Closure linker backend are useless at the
moment, because dynamic imports are only supported by the parser
and pretty-printer of GCC but not the rest of the compiler. Support
should eventually come in a later version of GCC, though, at which
point it will transfer for free to our GCC backend.