Recognize Metals as an IDE for the special treatment of the build.
The same accommodations that were done for Eclipse are also
valuable for Metals. Since the environment variable
`METALS_ENABLED` is automatically added by Metals when it imports
a build, this change gives a better out-of-the-box experience with
Metals.

We also now mention Metals-based IDEs as the recommended tools to
develop Scala.js in the Developing guide.