Implement `j.u.{Linked,}HashSet` on top of `j.u.{Linked,}HashMap`.
Previously, they were implemented on top of
`s.c.m.{Linked,}HashSet`, with hacks to support Java features and
behaviors, notably:

* Non-cooperative equality
* `Iterator.remove()`

Now that `java.util.{Linked,}HashMap`s are implemented from
scratch, it makes more sense to implement the hash sets on top of
them, since they already provide all these features in a native
way.

An even better way to implement them would be to do it from scratch
as well, with a copy of the algorithms for the hash maps, but the
only real advantage would be to claim a bit of memory usage. The
added complexity does not seem to be worth it.