Allow the Array.apply opt to apply to constant arrays of `j.l.Character`.
We remove the type parameters of `Array[Int](...)` so that the
*overload* for `Int` is selected, rather than the generic one,
which cannot be optimized for primitives.

Two arrays were `Array[Byte]`s, and this change turns them into
`Array[Int]`s, because `Array[Byte](...)` cannot be optimized
unless all elements are real bytes (not ints adaptable to bytes).
This is fine, though, because those two arrays were only used in
places were `Int`s are equally good, if not better.