Do not eagerly throw for `Nothing` in `Class.cast()`.
There is no need to, as the compiler will insert an explicit
cast at call site to adapt the erased `A` (`Object`) to `Nothing`
(assuming the call is not in statement position).

The test was a bit too restrictive, and has been relaxed not to
expect a specific error message, and to only expect an exception
when used in expression position.

The new behavior is consistent with Scala/JVM, since of course, on
the JVM, `Class.cast()` does not know about
`scala.runtime.Nothing$`.