Compile the javalanglib with -Yno-predef.
When writing code in the java.lang package, references to things
like `Boolean` or `Double` refer to `j.l.Boolean` or `j.l.Double`.
Usually this is not what we want (we want the primitive types
instead), but the implicits available in `Predef` hide mistakes by
introducing boxing and unboxing where required. The `-Yno-predef`
flag prevents these mistakes from happening.