Make `ConcurrentSkipListSet` independent of the internals of `TreeSet`.
In the JDK, `TreeSet.add()` specifically throws when adding the
first element, if it is not comparable to itself. This behavior
was introduced on purpose in JDK 7, so our `TreeSet` mimics that.

`ConcurrentSkipListSet`, however, allows the first element to be
added even if it is not comparable to itself. Since our
implementation of `ConcurrentSkipListSet` uses an inner `TreeSet`,
it would normally throw in that case. We had some code to
customize the inner `TreeSet` to cancel that throwing, but it was
relying on internal implementation details of `TreeSet`.

In this commit, we remove that specific code. This allows to make
`TreeSet.inner` private, hiding the fact that it uses an
`s.c.m.TreeSet` inside. However, this means that we relax a bit
the requirements of `ConcurrentSkipListSet`, as it will now throw
when trying to insert the first element if it is not comparable
to itself. The change in the tests highlights this change of
behavior.