Implement `java.util.TreeSet` from scratch.
Previously, it was implemented on top of `s.c.m.TreeSet`, with
various hacks to implement JDK features:

* `Iterator.remove()`
* Descending iterators and views
* Projections with exclusive start and/or inclusive end

All those hacks resulted in significantly more memory allocation
(at least one `Box` for most elementary operations) as well as
unreasonable time complexity for views.

The new implementation, done from scratch, has direct internal
support for all these operations:

* It never allocates during lookup and removal
* Iteration never allocates besides the instance of `Iterator`
* It allocates at most one object during `add`

The implementation was initially inspired by the implementation of
`s.c.m.TreeMap` in Scala 2.13.0 (in particular the core data
structure and algorithms in `RedBlackTree.scala`), but was heavily
changed to accommodate the requirements of the JDK API.

The red-black tree implementation supports values associated to
keys, although `TreeSet` does not use that feature. It will be
useful when we eventually implement `TreeMap`. The original
red-black tree in Scala 2.13.0 supported it, so it was easier to
port it than not.