Implement `java.util.PriorityQueue` from scratch.
Previously, it was implemented on top of `s.c.m.PriorityQueue`.
Unlike other collections implemented on top of Scala collections,
`PriorityQueue` had no serious hacks. It did have overhead in
terms of allocations of `Box`es, though.

The new implementation, done from scratch, does not box
elements, and offers direct support for all the methods of the
JDK API.

The test suite of `PriorityQueue` is enhanced in the process.
Notably, it includes all the standard tests for `Collection`s in
`CollectionTest`.