Do not override `j.l.Class` and `j.l.FloatingPointBits` in the minilib.
With the new independent javalanglib, the standard ones can now be
used in the minilib.

Moreover, we remove `scala.Serializable` from the minilib, as our
IR cleaner is able to remove its spurious addition by scalac. This
requires to adapt `RuntimeLong` a bit not to have the spurious
`scala.Serializable`, though (because it is not subject to the IR
cleaner).

We also adapt the standard `FloatPointBits` to directly use
`js.Math` methods instead of `j.l.Math`, so that we don't have to
add `Math` to the minilib.