Fix the spec of eq/ne for NaN and +0.0 v -0.0.
Previously, `x eq x` would answer `false` if `x` was `NaN`; and
`+0.0 eq -0.0` would answer `true`, although `+0.0.equals(-0.0)`
answers `false`.

That behavior was intentionally spec'ed, but arguably the spec had
been wrong all along. It was specified like that because we
couldn't possibly incur the performance hit of a strict identity
function for every single `eq` in the world, an operation that is
supposed to be virtually free.

Nevertheless, the spec'ed behavior caused a number of issues, for
example when using `eq` to cheaply test whether two values were
definitely equal (that was incorrect for +0.0 and -0.0), or when
using `eq` to find the exact occurrence of a value in a data
structure by reference (that was incorrect for NaN).

In this commit, we change the spec to the much more clearly correct
alternative where `NaN eq NaN` is true, and `+0.0 eq -0.0` is
false. The full check is implemented with the ES 2015 builtin
function `Object.is(x, y)`, which we polyfill when emitting ES 5.1.

In order not to have performance issues, the optimizer is enhanced
to optimize `eq` down to JavaScript's `===` when it can prove that
it is safe to do so. This happens when one of the following
applies:

* Either operand can be proven not to be a primitive number, which
  can often be done by looking at its type. This is a *very* common
  scenario, notably in all tests of the form `x eq null` or of the
  form `someNode eq x` (where `someNode` is a value of some type
  `Node` of a data structure, for example).
* Both operands are known to be whole primitive numbers: `x eq y`
  can become `x === y` when `x` and `y` are both `int`s, for
  example.

These optimizations leave very few uses of `eq` as the full check
with `Object.is`, all of which are actually desirable fixes. In
the test suite, the optimization applies to 99.8% of the use sites
of `eq` and `ne`.

This change means that there is no way to write user code
corresponding to ECMAScript's `x === y` for arbitrary `x` and `y`
anymore. Although it is possible to write it by hand on top of
`eq`, it is a bit absurd to undo the effects of `Object.is` to
recover the behavior of `===`. We add a new builtin
`js.special.strictEquals(x, y)` for that purpose. It is used in
the fast path of `BoxesRunTime.equals(x, y)`, which was the only
place in the codebase where we really meant `===` when writing
`eq`.