Upgrade the sbt plugin to sbt 1.2.1.
And test the sbt-plugin-test with sbt 1.2.8, which is the latest
stable release of sbt as of this writing.

The upstream fix https://github.com/sbt/zinc/pull/562, which made
its way into sbt 1.2.1, is necessary for incremental compilation
to even have a chance to work, as explained in #3419. However, it
is not enough, so an additional fix is still needed in our sbt
plugin.

The upgrade is not technically necessary for our build to succeed,
or even for the upcoming fix to build, but projects using Scala.js
will not benefit from the fix unless they upgrade to sbt 1.2.1+.
Depending on it in our build at least records the required version
in our published artifact.