Make AbstractCollection#toString match Java 8 specification.
The Java 8 specification for
(AbstractCollection)[https://docs.oracle.com/
javase/8/docs/api/java/util/AbstractCollection.html#toString--]
describes the resultant string as using comma-space as the separator for
collections with size greater than one. Prior to this PR, ArrayDeque.scala,
ArrayList, and other sub-classes of AbstractCollection would create a
string such as "[a,b]". This PR corrects that to the specified "[a, b]".