Separate definitions of JS members into their own IR nodes ...
... and get rid of `PropertyName`s, replacing them by either `Ident`
or `Tree`, as appropriate.

Previously, the `name` field of `FieldDef`s, `MethodDef`s and
`PropertyDef`s was a `PropertyName`, which could be one of three
things: an `Ident`, a `StringLiteral` or a `ComputedName`. If it
was an `Ident`, the member was a statically typed Scala member;
whereas if it was a `StringLiteral` or `ComputedName`, it was a
dynamically typed JS member.

Very different rules apply to Scala members versus JS members,
from their reachability analysis to their IR checking to their
emission to JavaScript code. Distinguishing the code paths based
on the type of `name` was awkward.

In this commit, we separate Scala and JS member definitions into
specific IR nodes: `FieldDef` v `JSFieldDef`, `MethodDef` v
`JSMethodDef`, and `JSPropertyDef` (there are no Scala property
definitions).

Scala member definitions always have a `name: Ident`, while JS
members always have a `name: Tree`. `StringLiteral`s and
`ComputedName`s are merged in a single case of `Tree`.

The replacement of `ComputedName` by `Tree` means that the
`logicalName` does not exist anymore. Logical names were hacky and
not well specified, so this removes a quirk of the IR. However,
some aspects of the linker had to be adapted to compensate:

* The `Infos` for the reachability analysis now have a dedicated
  slot for the reachability info of exported members. The
  `Analyzer` is adapted to deal with that.
* In the `Refiner`, the caching mechanism for the `Infos` of
  exported members now uses their index in the list of exported
  members as their identity. (More details available in `Refiner`).

As a by-product of removing logical names, this commit fixes #3784,
since the IR checker can let through property names containing
`__`.