[no-master] Fix #3778: Never inline `clone__O` if the receiver can be an array.
Backport of 56371e91ee38b0c9cd262803351b2e8c2baa0cff.

The `clone__O()` method of arrays is very peculiar in our pipeline:
it is the only method that is defined in array classes, for which
no IR exists. There are several pieces of the linker that special-
case `clone__O()` for arrays for that reason, but we were missing
one detail: the optimizer must never try to inline `clone__O()` if
the receiver can be an array.

Before this commit, the optimizer could erroneously do that,
because the Class Hierarchy Analysis done by `GenIncOptimizer` does
not have any provision for array classes, and would return the only
reachable `clone__O()` method if there was one in the classpath (on
`Object` or on an arbitrary other class).

In this commit, we specifically forbid inlining of `clone__O()` if
the receiver can be an array.

Since the test suite reaches the `clone__O()` method of many
classes, it is impossible to reproduce the problematic behavior in
the test suite. The original fix in 1.x instead uses linker unit
tests that make assertions about the IR produced by the optimizer.
This backport does not include tests because the 0.6.x branch does
not have the necessary infrastructure.