Expose ModuleInitializer to users of standard
This is necessary if we are to split the interface from the
implementation. Note that it also makes the standard linker less
special, as now any linker can deal with module initializers.

Before this commit it was necessary to use some parts of the standard
linker to implement the Linker interface.