[no-master] Backport `js.special.strictEquals`.
Since the definition of `x eq y` has changed in Scala.js 1.x, it
might be necessary to use `js.special.strictEquals` where
previously `eq` was correct. This can make cross-compilation a bit
difficult.

This commit backports the definition of `js.special.strictEquals`
to 0.6.x, with a trivial implementation that delegates to `eq`.
Tests are directly imported from 1.x.