Use a `Type` rather than a `TypeRef` in `Is/AsInstanceOf`, ...
... and fuse `Unbox` into `AsInstanceOf`.

The use of a `TypeRef` was a historical artifact of when `Type`s
did not have separate cases for the "int types" (i.e., `byte`,
`short`, `int` and `char`). Since they can now represent all the
primitive types separately, it is more appropriate to just use a
`Type`.

This also fits the fact that type-refs to JS types are not valid
in these nodes, just like they are not valid `Type`s.

And since we can represent primitive types in `AsInstanceOf`, we
can fuse `Unbox` into it: an `AsInstanceOf` with a primitive type
is an `Unbox`.

For symmetry, we add support for primitive types in `IsInstanceOf`
nodes as well. For those, they are in fact equivalent to testing
for the corresponding boxed classes. We use the support for
primitive types in `IsInstanceOf` nodes in the overload resolution
of exports.