Introduce dedicated types for the kinds of `Ident`s and their `Name`s.
Instead of putting everything in generic `Ident`s with `String`s,
we now have dedicated types a) of `Name`s and b) of `Ident`s for
the following categories:

* Local variables, including parameters and captures
* Labels
* Fields
* Methods
* Classes

The `Name`s such as `LocalName` and `MethodName` are currently
implemented as bare `String`s hidden behind a type alias, and
protected by validation upon construction. They are all validated
to be valid JS identifiers, which was what they already had to be
before this commit.

In the future, this new abstraction will make it easier to
introduce different rules for validation, and in particular to
delay mangling to the emitter, and to have better structure for
method names (whose signatures are meaningful).