Fix #2108: Defer identifier name mangling to the emitter.
We introduce specialized classes for Names in the IR, instead of
using aliases of Strings. The new Names wrap an `EncodedName`,
which is an `Array[Byte]` for the UTF-8 encoding of a name. They
are optimized for frequent comparisons and hash-code access.

Simple names must be non-empty and can contain all Unicode code
points, with the exception of / ; . [

Local names, label names and field names are just simple names.
Class names are non-empty sequences of simple names separated by
dots (`.`). The simple name of method names must not contain the
code point `<` unless they are exactly `<init>` or `<clinit>`.

The above restrictions exactly match what the JVM allows, so our
scalac back-end can directly translate all symbol names to the
IR without further mangling, just like the JVM back-end.

`MethodName`s are not just the *simple* name of a method. They
also encode the full signature, where param type refs and result
type ref are kept as is. Therefore, `MethodName`s do not directly
have an `EncodedName` representing them entirely.

---

The responsibility of mangling those arbitrary Unicode strings into
JavaScript identifiers is delegated to the emitter, in particular
in `JSGen`. The encoding is as follows:

* ASCII characters that are valid in JS identifiers are kept as is.
* The individual *bytes* of the UTF-8 encoding that are not valid
  ASCII characters are translated into `(b + 256).toChar`. It turns
  out that all code points in the range [256,512) are valid in JS
  identifiers (including as first char).
* For local names, a leading `$` is rewritten as `\u03b4` (greek
  small letter delta) so that it does not clash with "envFields".
* For class names, `.` is rewritten as `_` and `_` is rewritten as
  `\uff3f` (FULLWIDTH SPACING UNDERSCORE).
* For method names, param type refs and result type refs are
  encoded using the `I` or `Llong.Name` scheme, and separated with
  `__`, which is basically how they were encoded before.

As before, a number of common package names are compressed as short
prefixes. The classes `java.lang.Object` and `java.lang.String` are
compressed as `O` and `T`, respectively. This applies both to whole
class names and within signatures.

---

A lot of code in this commit has been written with raw efficiency
in mind. Names are manipulated very often, so it is essential that
operations on them be as fast as possible.