Hide private fields of anonymous JS classes.
They are hidden in a separate object, which is itself stored in a
non-enumerable field of the anonymous object. When possible, that
field has a `Symbol` name. Otherwise, it is a string that is
randomly generated at run-time with the same enthropy as a UUID.

Fixes #2748 because the non-enumerable field does not show up in
`JSON.stringify` nor other kinds of cursory inspections (it can be
found with `getOwnPropertySymbols()` and
`getOwnPropertyDescriptors()`).

Fixes #3777 because we get rid of the `FieldDef`s that were
declared in abstract JS types. The IR checker checks anew that
abstract JS types do not have field declarations.