Fix #2971: Ensure that `FunctionEmitter`'s temp vars don't clash.
First, we refactor `newSyntheticVar()` to return an already
transformed `js.Ident` rather than a `LocalIdent`. This has
several consequences, notably the fact that we need to temporarily
manipulate IR `Tree`s that contain refereces to `js.Ident`
variables. We do this with a `Transient` node. The refactoring
also simplifies the handling of records in the emitter.

Once that is done, we trivially change `newSyntheticVar()` to use
"envFields" for temporary vars, as they already have the necessary
mechanisms not to clash with `LocalIdent`s nor with global refs.