Fix #3822: Protect inlined global refs in `typeof`.
By spec, if the direct argument of `js.typeOf()` is a global ref,
it must evalute to `"undefined"` if the global ref is not defined.
That was correctly implemented, but it also leaked into cases where
the argument was not a global ref at compile-time, but became one
after optimizations (typically inlining).

We now introduce a dedicated IR node for the magic `typeof` of a
global ref that must return `"undefined"`, which we emit at
compile-time where appropriate. The normal `typeof` protects its
argument if it ends up being a global ref.