Fix #2680: Use Symbols for the private fields of JS classes.
Any encoding of field names as identifiers that we could come up
would inevitably be prone to clashes, as we can never know which
fields would be declared in superclasses, or worse, in subclasses.

Therefore, we now create one `Symbol` for each private field
declaration, and we use that. When Symbols are not supported by the
engine, we dynamically generate random names that have the same
entropy as UUIDs. This does not guarantee that there will never be
any clash, but it makes the likelihood close to zero.

There is no additional test, because there is no way to write a
test that would have failed before this commit, and would still
be relevant after this commit, given that the compilation has
completely changed. Even removing the constraint that a test would
need to fail before the commit, I cannot think of any test that
would be relevant for this commit.