Check that methods called exist on the static type, even if abstract.
Previously, the following program would link:

    class A
    class B {
      def foo(): Int = 1
    }
    val a: A = new B()
    a.foo()

even though `foo` is not declared in `A`, because `A` itself is
never instantiated.

This is not a problem for the current JavaScript emitter, nor for
the optimizer, but it can be problematic in the future for:

* IR-checking that method calls are correct (#1419)
* a potential Wasm back-end, which might want to rely on the
  abstract methods being declared to build virtual method tables.

This commit performs the bare minimum that needs to be done before
1.0.0 in order to future-proof those two things: in the base linker
(only), check that methods called exist on the static type of the
argument, potentially as abstract methods.