Directly use `ClassName`s instead of `ClassRef`s in the IR `Tree`s.
It made sense to use `ClassRef`s when they wrapped bare `String`s,
as it provided better type safety. However, now that we have
type-safe `ClassName`s, the wrapping `ClassRef` are just noise and
useless indirections.

We also more consistently use `className` as a variable of type
`ClassName`, instead of `cls` which was used for `ClassRef`s.

After this commit, `ClassRef`s are only used as one case of
`TypeRef`, as they should.