Fix #3830: Make isReflectiveProxy an explicit flag of MethodName.
In doing so, we can turn `resultTypeRef` from being optional to
being mandatory. For constructors and static initializers, it must
be set to `void`, whereas for reflective proxies, it must be set
to `java.lang.Object`. This is consistent with how they are
typechecked both at definition site and call site.