Make `Name`s safer wrt. mutations and invalid states.
Instead of being a raw `Array[Byte]`, we now use a dedicated
class `UTF8String` which is an immutable valid UTF-8 sequence,
validated as such when constructed. This allows to expose it to
users of `Name`s in a safe way.

We also remove or hide all the constructors that by-pass name
validation, so that it is not possible to create invalid names
from outside the `ir` package.