Fix #3839: Treat objects/classes nested in `@JSGlobalScope` as global refs.
An `@js.native` object or class that is nested in an
`@JSGlobalScope` object must be treated as a definition for a
global ref. It is very similar to being annotated with `@JSGlobal`,
except that path selections are not accepted, since
`@JSName("foo.bar")` is supposed to mean selecting a member named
`foo.bar`, whereas `@JSGlobal("foo.bar")` is supposed to mean
selecting the global ref `foo`, and inside it select the member
named `bar`.