Encode original names as optional `UTF8String`s.
We add a dedicated value class `OriginalName` for an optional
`UTF8String` intended to be used as the original name of an
identifier.

This avoids decoding `Name`s when we infer an original name from a
previous `Name`. This specific operation is even directly offered
as `orElse(Name)` (and variants).

This change means that original names must now always be
well-formed Unicode strings. Unpaired surrogates are not valid.