Store original names only on definitions, not on every usage.
The original name at the use site of identifiers should always
match the original name at its definition site. Therefore, it is
wasteful to carry around the original names at use sites in the IR,
both in terms of processing time and memory usage.

In this commit, we move the `originalName: OriginalName`
parameters from the `xIdent`s to the `xDef`s. The `Ident`s now only
carry a `Name` and a `pos: Position`.

We use the forced refactoring to make sure everywhere that we
compute appropriate `OriginalName`s when assigning fresh names to
things: when creating a fresh name `fresh` from a `base` name with
an `originalName`, the new original name becomes:

* `originalName` if it is defined,
* `NoOrignalName` if `fresh eq base`,
* `OriginalName(base)` otherwise.

When generating the JavaScript trees, we currently do not carry
original names from definition site to use site, which means we
never assign original names to use sites. This is different than
what we used to do before this commit. This is, in theory, a
regression, but since there is no real specification of what
original names in source maps are supposed to mean, and since I
could not experience any observable regression with source maps,
this is probably fine. We can rearrange this in a binary compatible
way in the future if we find out that this is not the right thing
to do.