Fix #3840: Rename `envField` to `codegenVar`.
`envField` was a historical name that can traced back to when all
compiler-generated variables were stored as fields of a global
object `ScalaJS`. The name was not too bad until recently, when we
started using the mechanism even for some local variables.

This commit renames all "env field" occurrences to "codegen var",
since those are essentially codegen-chosen variable names.