Fix #3888: Always identify impl classes by their IMPLCLASS flag.
Before, we sometimes used name-based identification of impl
classes. This was introduced in
ac54f189d225c7c66a7156f3b41594ee2f43a01d, before Scala.js 0.1! At
the time, we were still using a custom `Global` subtrait instead of
a compiler plugin, and a custom `.jstype` file reader instead of
the normal class-file reader of scalac. The flag was not present in
some cases of separate compilation. Now that we use a normal
compiler plugin, this issue does not seem to manifest itself
anymore. It seems this code was left there for 6 years for nothing.