Do not special-case implementation classes so much.
Now that we have proper namespaces for methods, including static
namespaces, it is better to generate static definitions and static
calls based on `sym.isStaticMember` rather than
`sym.owner.isImplClass`. This removes some special cases for impl
classes, and future-proofs our back-end in case scalac decides to
emit more stuff as static (e.g., when `this` is not used in a
lambda, not to unnecessarily retain the reference to `this`).