Generate static forwarders, and use them to call Java-defined stuff.
Previously, when calling a Java-defined static method, we knew that
it did not actually exist as such, but rather was defined as an
instance method of the companion module class. Therefore, we added
a redirection at call-site.

The fact that static methods of Java APIs were not actually static
was basically part of the IR specification for the javalib, which
was not nice at all.

In this commit, we start generating static forwarders in the class
or interface, that call to the module class' instance methods. At
call site, when calling a Java-defined static method, we now do
exactly as Scala/JVM does, i.e., we actually call a static method.

The optimizer gets rid of all those forwarders, since they are all
trivial.