Make `binaryCross` version same as major for minor milestones.
For example, in version 1.2.0-M1 that emits IR v1.2-M1, the binary
cross version is now `1` instead of `1.2-M1`. This makes the
behavior between snapshots and milestones uniform.

More importantly, this change allows to always derive the binary
cross from the full Scala.js version, using 2 simple rules:

* If the version is `x.y.z-w`, then the binary cross is `x.y-w`.
* If the version is `x.y.z`, then the binary cross is `x`.

Before this change, in a situation where the full Scala.js version
is a milestone (e.g., 1.3.0-M1), it was not possible to derive the
binary cross version from it. It might be 1.3-M1 if the binary
emitted version was 1.3-M1, but it might also be 1 if the binary
emitted version was 1.2, for example.

Some build tools do not have access to the binary emitted version;
they only have access to the full Scala.js version. It is important
for them to be able to derive the binary cross version from the
full version only.